+++ src/timeline.js
@@ -60,7 +60,9 @@ const testids = createTestids('Timeline', {
   selector: '',
   table: '',
   ganttBody: '',
-  splitPaneResizer: ''
+  splitPaneResizer: '',
+  fadeEffect: '',
+  timeBar: ''
 });
 export const timelineTestids = testids;
 
@@ -79,6 +81,11 @@ export const DEFAULT_ROW_ODD_CLASS = '';
 export const DRAG_TO_CREATE_POPUP_CLOSE_TIME = 5000;
 export const DRAG_TO_CREATE_POPUP_LABEL_2 = 'Popup will close in a few moments.';
 
+const FADE_OPACITY_OFFSET = 0.1;
+const FADE_TIMER_INTERVAR = 100;
+export const ZOOM_PERCENT = 0.2;
+export const MIN_DISPLAY_TIME = 60000;
+
 export const PARENT_ELEMENT = componentId => document.querySelector(`.rct9k-id-${componentId} .parent-div`);
 
 const TableWithStyle = ({table}) => {
@@ -96,6 +103,8 @@ const TableWithStyle = ({table}) => {
 };
 
 export const DRAG_TO_CREATE_ACTION_LABEL = 'Drag to create';
+export const ZOOM_IN_ACTION_LABEL = 'Zoom in';
+export const ZOOM_OUT_ACTION_LABEL = 'Zoom out';
 /**
  * Timeline class
  * @extends React.Component<Timeline.propTypes>
@@ -460,9 +469,34 @@ export default class Timeline extends React.Component {
      * If this handler is provided, it will be called when the table is resized
      * by dragging the split bar between the table and the gantt.
      *
-     * @type {(tableWidth: number) => void}
+     * @type {(splitSize: number) => void}
      */
-    onTableResize: PropTypes.func
+    onSplitChange: PropTypes.func,
+
+    /**
+     * This property controls if one segment (item) whose period overlaps another segment (item)'s period
+     * is displayed on a separate subrow or it displayes on the same subrow, overlapping the other segment
+     *
+     * In case of overlapping segments displayed on separate subrows, the main row expands its height to include all the subrows
+     *
+     * @default true
+     * @type { boolean | (item: Item, rowIndex: number) => boolean }
+     */
+    displayItemOnSeparateRowIfOverlap: PropTypes.oneOfType([PropTypes.bool, PropTypes.func]),
+
+    /**
+     * The segments with bigger index are staying in front of the ones with smaller index
+     *
+     * @type {(item: Item) => number}
+     */
+    zIndexFunction: PropTypes.func,
+
+    /**
+     * If this property is provided with true, then in context menu displays the shortcut for zoom in and zoom out.
+     *
+     * @type { undefined | boolean}
+     */
+    showZommShortcuts: PropTypes.bool
   };
 
   static defaultProps = {
@@ -481,7 +515,6 @@ export default class Timeline extends React.Component {
     onItemHover() {},
     onItemLeave() {},
     interactOptions: {},
-    itemStyle: {},
     rowClassName: DEFAULT_ROW_CLASS,
     rowEvenClassName: DEFAULT_ROW_EVEN_CLASS,
     rowOddClassName: DEFAULT_ROW_ODD_CLASS,
@@ -491,7 +524,7 @@ export default class Timeline extends React.Component {
     minDate: undefined,
     maxDate: undefined,
     selectedItems: undefined,
-    snap: undefined,
+    snap: 1,
     timebarFormat: undefined,
     bottomResolution: undefined,
     topResolution: undefined,
@@ -509,7 +542,12 @@ export default class Timeline extends React.Component {
     onDragToCreateEnded: undefined,
     onContextMenuShow: undefined,
     onSelectionChange() {},
-    onTableResize: undefined
+    onSplitChange: undefined,
+    displayItemOnSeparateRowIfOverlap: true,
+    zIndexFunction() {
+      return 3;
+    },
+    showZommShortcuts: false
   };
 
   /**
@@ -539,6 +577,11 @@ export default class Timeline extends React.Component {
    */
   static no_op = () => {};
 
+  /**
+   * Fade timer
+   */
+  fadeTimer = undefined;
+
   getInitialTableWidth(props) {
     props = props ? props : this.props;
     return (props.table ? props.table.props.width : 0) + TABLE_OFFSET;
@@ -547,7 +590,6 @@ export default class Timeline extends React.Component {
   constructor(props) {
     super(props);
     this.selecting = false;
-    this.getInitialTableWidth = this.getInitialTableWidth.bind(this);
     this.state = {
       selection: [],
       cursorTime: null,
@@ -555,7 +597,7 @@ export default class Timeline extends React.Component {
       verticalGridLines: [],
       screenHeight: 0,
       gridWidth: 0,
-      tableWidth: this.getInitialTableWidth(),
+      splitSize: props.table ? props.table.props.width : 0,
       dragToCreateMode: false,
       dragToCreatePopupClosed: false,
       openMenu: false,
@@ -569,6 +611,9 @@ export default class Timeline extends React.Component {
       endDate: this.props.endDate,
       hasHorizontalScrollbar: false,
       touchPositionX: undefined,
+      fadeEffectOpen: false,
+      fadeEffectContent: undefined,
+      fadeEffectOpacity: 0,
       illegalInputData: false
     };
 
@@ -612,8 +657,11 @@ export default class Timeline extends React.Component {
     this.setVerticalGridLines = this.setVerticalGridLines.bind(this);
     this.handleScrollTable = this.handleScrollTable.bind(this);
     this.handleScrollGantt = this.handleScrollGantt.bind(this);
-    this.handleDrag = this.handleDrag.bind(this);
+    this.onSplitChange = this.onSplitChange.bind(this);
     this.selectionChangedHandler = this.selectionChangedHandler.bind(this);
+    this.getRowClassName = this.getRowClassName.bind(this);
+    this.wheelHandler = this.wheelHandler.bind(this);
+    this.startFadeOutEffect = this.startFadeOutEffect.bind(this);
 
     const canSelect = Timeline.isBitSet(Timeline.TIMELINE_MODES.SELECT, this.props.timelineMode);
     const canDrag = Timeline.isBitSet(Timeline.TIMELINE_MODES.DRAG, this.props.timelineMode);
@@ -623,6 +671,7 @@ export default class Timeline extends React.Component {
 
   componentDidMount() {
     window.addEventListener('resize', this.updateDimensions);
+    window.addEventListener('wheel', this.wheelHandler, {passive: false});
   }
 
   componentWillReceiveProps(nextProps) {
@@ -645,10 +694,6 @@ export default class Timeline extends React.Component {
       }
     }
 
-    if (this.getInitialTableWidth(nextProps) !== this.getInitialTableWidth(this.props)) {
-      this.setState({tableWidth: this.getInitialTableWidth(nextProps)});
-    }
-
     if (this.props.startDate != nextProps.startDate || this.props.endDate != nextProps.endDate) {
       this.setState({startDate: nextProps.startDate, endDate: nextProps.endDate});
     } else {
@@ -656,7 +701,8 @@ export default class Timeline extends React.Component {
         nextProps.items,
         convertDateToMoment(nextProps.startDate, nextProps.useMoment),
         convertDateToMoment(nextProps.endDate, nextProps.useMoment),
-        nextProps.useMoment
+        nextProps.useMoment,
+        nextProps.displayItemOnSeparateRowIfOverlap
       );
     }
     this.fillInTimelineWithEmptyRows(nextProps.groups);
@@ -672,6 +718,7 @@ export default class Timeline extends React.Component {
     if (this._selectRectangleInteractable) this._selectRectangleInteractable.unset();
 
     window.removeEventListener('resize', this.updateDimensions);
+    window.removeEventListener('wheel', this.wheelHandler);
   }
 
   componentDidUpdate(prevProps, prevState) {
@@ -694,6 +741,48 @@ export default class Timeline extends React.Component {
         this.props.useMoment
       );
       this.refreshGrid();
+    }
+
+    if (prevState.hasHorizontalScrollbar != this.state.hasHorizontalScrollbar) {
+      this.fillInTimelineWithEmptyRows(this.props.groups);
+      this.refreshGrid();
+    }
+  }
+
+  wheelHandler(e) {
+    if (e.ctrlKey) {
+      let target = e.target;
+      while (target) {
+        if (target.className.includes('rct9k-grid')) {
+          break;
+        }
+        target = target.parentElement;
+      }
+      if (!target) {
+        return;
+      }
+      e.preventDefault();
+      e.stopPropagation();
+
+      const interval = moment(this.state.endDate).valueOf() - moment(this.state.startDate).valueOf();
+      const delta = e.clientX / this._grid.props.width;
+      let deltaInterval = interval * ZOOM_PERCENT;
+      if (e.deltaY > 0) {
+        deltaInterval *= -1;
+      }
+      let startDate = moment(moment(this.state.startDate).valueOf() + delta * deltaInterval);
+      let endDate = moment(moment(this.state.endDate).valueOf() - (1 - delta) * deltaInterval);
+      if (endDate - startDate < MIN_DISPLAY_TIME) {
+        return;
+      }
+      if (startDate.valueOf() < moment(this.getMinDate()).valueOf()) {
+        startDate = this.getMinDate();
+      }
+      if (endDate.valueOf() > moment(this.getMaxDate()).valueOf()) {
+        endDate = this.getMaxDate();
+      }
+      this.setState({startDate, endDate});
+      this.throttledMouseMoveFunc(e);
     }
   }
 
@@ -734,7 +823,7 @@ export default class Timeline extends React.Component {
     if (this.props.minDate) {
       return convertDateToMoment(this.props.minDate, this.props.useMoment);
     } else {
-      return this.getStartDate();
+      return convertDateToMoment(this.props.startDate, this.props.useMoment);
     }
   }
 
@@ -747,7 +836,7 @@ export default class Timeline extends React.Component {
     if (this.props.maxDate) {
       return convertDateToMoment(this.props.maxDate, this.props.useMoment);
     } else {
-      return this.getEndDate();
+      return convertDateToMoment(this.props.endDate, this.props.useMoment);
     }
   }
 
@@ -842,6 +931,20 @@ export default class Timeline extends React.Component {
   }
 
   /**
+   * Before the gantt leftOffset was considered to be: PARENT_ELEMENT.getBoundingClientRect().left
+   * But there was a bug found in tests: The segments are in fact positioned (in the html page) relative to the interior of the react virtualized grid
+   * that is smaller than the .parent-div with 1px (Because of a 1px left border of the grid)
+   *
+   * So the left offset of the gantt (needed when converting the position in pixels in time) should take in consideration
+   * that 1px left border of the gantt grid
+   *
+   * @return {number}
+   */
+  getGanttLeftOffset() {
+    return PARENT_ELEMENT(this.props.componentId).getBoundingClientRect().left + 1;
+  }
+
+  /**
    * Re-renders the grid when the window or container is resized
    */
   updateDimensions() {
@@ -860,11 +963,17 @@ export default class Timeline extends React.Component {
    * @param {boolean} useMoment This parameter is necessary because this method is also called when
    * the component receives new props.
    */
-  setTimeMap(items, startDate, endDate, useMoment) {
+  setTimeMap(items, startDate, endDate, useMoment, displayItemOnSeparateRowIfOverlap) {
     if (!startDate || !endDate) {
       startDate = this.getStartDate();
       endDate = this.getEndDate();
     }
+
+    displayItemOnSeparateRowIfOverlap =
+      displayItemOnSeparateRowIfOverlap == undefined
+        ? this.props.displayItemOnSeparateRowIfOverlap
+        : displayItemOnSeparateRowIfOverlap;
+
     this.itemRowMap = {}; // timeline elements (key) => (rowNo).
     this.rowItemMap = {}; // (rowNo) => timeline elements
     this.rowHeightCache = {}; // (rowNo) => max number of stacked items
@@ -893,7 +1002,9 @@ export default class Timeline extends React.Component {
         maxVisibleItems,
         this.getStartFromItem,
         this.getEndFromItem,
-        useMoment
+        useMoment,
+        displayItemOnSeparateRowIfOverlap,
+        rowInt
       );
     });
   }
@@ -1032,6 +1143,12 @@ export default class Timeline extends React.Component {
       return this.props.snapMinutes * 60 * 1000;
     }
     return 1;
+  }
+
+  getRowClassName(rowIndex) {
+    return (
+      this.props.rowClassName + ' ' + (rowIndex % 2 == 0 ? this.props.rowOddClassName : this.props.rowEvenClassName)
+    );
   }
 
   /**
@@ -1204,10 +1321,9 @@ export default class Timeline extends React.Component {
       if (event.button == 2) {
         this.setState({openedContextMenuCoordinates: {x: event.clientX, y: event.clientY}});
         this.setState({openedContextMenuRow: bottomRow});
-        const leftOffset = PARENT_ELEMENT(this.props.componentId).getBoundingClientRect().left;
         this.setState({
           openedContextMenuTime: getTimeAtPixel(
-            event.clientX - leftOffset,
+            event.clientX - this.getGanttLeftOffset(),
             this.getStartDate(),
             this.getEndDate(),
             this.getTimelineWidth(),
@@ -1473,7 +1589,9 @@ export default class Timeline extends React.Component {
             let new_row_height = getMaxOverlappingItems(
               this.rowItemMap[rowNo],
               this.getStartFromItem,
-              this.getEndFromItem
+              this.getEndFromItem,
+              this.props.displayItemOnSeparateRowIfOverlap,
+              rowNo
             );
             if (new_row_height !== this.rowHeightCache[rowNo]) {
               this.rowHeightCache[rowNo] = new_row_height;
@@ -1549,9 +1667,8 @@ export default class Timeline extends React.Component {
       }
     } else {
       row = e.target.getAttribute('data-row-index');
-      const leftOffset = PARENT_ELEMENT(this.props.componentId).getBoundingClientRect().left;
       let clickedTime = getTimeAtPixel(
-        e.clientX - leftOffset,
+        e.clientX - this.getGanttLeftOffset(),
         this.getStartDate(),
         this.getEndDate(),
         this.getTimelineWidth()
@@ -1569,10 +1686,9 @@ export default class Timeline extends React.Component {
       // right click => open CM
       this.setState({openedContextMenuCoordinates: {x: e.clientX, y: e.clientY}});
       this.setState({openedContextMenuRow: Number(row)});
-      const leftOffset = PARENT_ELEMENT(this.props.componentId).getBoundingClientRect().left;
       this.setState({
         openedContextMenuTime: getTimeAtPixel(
-          e.clientX - leftOffset,
+          e.clientX - this.getGanttLeftOffset(),
           this.getStartDate(),
           this.getEndDate(),
           this.getTimelineWidth(),
@@ -1619,11 +1735,7 @@ export default class Timeline extends React.Component {
           key={key}
           style={style}
           data-row-index={rowIndex}
-          className={
-            this.props.rowClassName +
-            ' ' +
-            (rowIndex % 2 == 0 ? this.props.rowOddClassName : this.props.rowEvenClassName)
-          }
+          className={this.getRowClassName(rowIndex)}
           onClick={e => this._handleItemRowEvent(e, Timeline.no_op, this.props.onRowClick)}
           onMouseDown={e => {
             this.selecting = false;
@@ -1651,7 +1763,10 @@ export default class Timeline extends React.Component {
             this.props.itemRendererDefaultProps,
             this.getStartFromItem,
             this.getEndFromItem,
-            timelineTestids
+            timelineTestids,
+            this.props.displayItemOnSeparateRowIfOverlap,
+            this.props.zIndexFunction,
+            rowIndex
           )}
           {rowLayerRenderer(
             layersInRow,
@@ -1764,9 +1879,8 @@ export default class Timeline extends React.Component {
    * Only calls back if a new snap time is reached
    */
   throttledMouseMoveFunc(e) {
-    const leftOffset = PARENT_ELEMENT(this.props.componentId).getBoundingClientRect().left;
     const cursorSnappedTime = getTimeAtPixel(
-      e.clientX - leftOffset + 2,
+      e.clientX - this.getGanttLeftOffset(),
       this.getStartDate(),
       this.getEndDate(),
       this.getTimelineWidth(),
@@ -1898,10 +2012,11 @@ export default class Timeline extends React.Component {
     return true;
   };
 
-  handleDrag(width) {
-    this.setState({tableWidth: width});
-    if (this.props.onTableResize) {
-      this.props.onTableResize(width);
+  onSplitChange(width) {
+    if (this.props.onSplitChange) {
+      this.props.onSplitChange(width);
+    } else {
+      this.setState({splitSize: width});
     }
     return true;
   }
@@ -2012,6 +2127,41 @@ export default class Timeline extends React.Component {
         }
       });
     }
+    if (this.props.showZommShortcuts) {
+      let that = this;
+      actions.push({
+        label: ZOOM_IN_ACTION_LABEL,
+        icon: 'zoom-in',
+        run: param => {
+          let event = new MouseEvent('wheel', {
+            ctrlKey: true,
+            clientX: that._gridDomNode.getBoundingClientRect().x + that._grid.props.width / 2,
+            bubbles: true,
+            cancelable: true
+          });
+          event.deltaY = -1;
+          that._gridDomNode.dispatchEvent(event);
+          param.closeContextMenu();
+          that.startFadeOutEffect('Zommed in');
+        }
+      });
+      actions.push({
+        label: ZOOM_OUT_ACTION_LABEL,
+        icon: 'zoom-out',
+        run: param => {
+          let event = new MouseEvent('wheel', {
+            ctrlKey: true,
+            clientX: that._gridDomNode.getBoundingClientRect().x + that._grid.props.width / 2,
+            bubbles: true,
+            cancelable: true
+          });
+          event.deltaY = 1;
+          that._gridDomNode.dispatchEvent(event);
+          param.closeContextMenu();
+          that.startFadeOutEffect('Zommed out');
+        }
+      });
+    }
 
     return (
       <ContextMenu
@@ -2019,6 +2169,36 @@ export default class Timeline extends React.Component {
         positionToOpen={actions.length > 0 ? this.state.openedContextMenuCoordinates : undefined}
         actions={actions}
       />
+    );
+  }
+
+  /**
+   * Show a popup with fadeEffectContent, and start the fadeOut
+   * @param { string | JSX.Element } fadeEffectContent
+   */
+  startFadeOutEffect(fadeEffectContent) {
+    this.setState({fadeEffectOpen: true, fadeEffectContent, fadeEffectOpacity: 1});
+    let that = this;
+    this.fadeTimer = setInterval(function() {
+      if (that.state.fadeEffectOpacity > 0) {
+        that.setState({fadeEffectOpacity: that.state.fadeEffectOpacity - FADE_OPACITY_OFFSET});
+      } else {
+        clearInterval(that.fadeTimer);
+      }
+    }, FADE_TIMER_INTERVAR);
+  }
+
+  /**
+   * @returns { JSX.Element }
+   */
+  renderFadeEffect() {
+    return (
+      <Popup
+        data-testid={testids.fadeEffect}
+        open={this.state.fadeEffectOpen}
+        style={{opacity: this.state.fadeEffectOpacity}}>
+        {this.state.fadeEffectContent}
+      </Popup>
     );
   }
 
@@ -2151,6 +2331,7 @@ export default class Timeline extends React.Component {
                       this.setState({hasHorizontalScrollbar: isScrollbarVisible})
                     }
                     ref={node => (this._scrollbar = node)}></Scrollbar>
+                  {this.renderFadeEffect()}
                   {this.renderContextMenu()}
                   {backgroundLayer &&
                     React.cloneElement(backgroundLayer, {
@@ -2258,8 +2439,8 @@ export default class Timeline extends React.Component {
                   <SplitPane
                     split="vertical"
                     style={{height: this.state.screenHeight, position: 'relative'}}
-                    size={this.state.tableWidth}
-                    onChange={this.handleDrag}
+                    size={this.props.onSplitChange ? this.props.table.props.width : this.state.splitSize}
+                    onChange={this.onSplitChange}
                     ref={this.splitPane_ref_callback}>
                     <TableWithStyle
                       table={React.cloneElement(this.props.table, {
@@ -2272,14 +2453,21 @@ export default class Timeline extends React.Component {
                         scrollTop: this.state.scrollTop,
                         headerHeight: timebarHeight,
                         height: this.state.screenHeight,
-                        width: this.state.tableWidth,
-                        rowClassNameGetter: rowIndex =>
-                          this.props.rowClassName +
-                          ' ' +
-                          (rowIndex % 2 == 0 ? this.props.rowOddClassName : this.props.rowEvenClassName),
+                        rowClassNameGetter: rowIndex => {
+                          let classNameDefinedByUser =
+                            this.props.table !== undefined &&
+                            this.props.table.props.rowClassNameGetter !== undefined &&
+                            this.props.table.props.rowClassNameGetter(rowIndex);
+                          if (classNameDefinedByUser) {
+                            return classNameDefinedByUser;
+                          } else {
+                            return this.getRowClassName(rowIndex);
+                          }
+                        },
                         // Because the content of the empty rows are now significant
                         // avoid showing vertical scrollbar in case horizontal scrollbar is needed when shrinking the table
-                        showScrollbarY: !hasEmptyRows
+                        showScrollbarY: !hasEmptyRows,
+                        ...(this.props.onSplitChange ? {} : {width: this.state.splitSize})
                       })}
                     />
                     {this.renderGanttPart({bodyHeight, timebarHeight})}
@@ -2306,9 +2494,8 @@ export default class Timeline extends React.Component {
   ////////////////////////////////////////////////////////////////////////////////////////
 
   setCursorTime(x) {
-    const leftOffset = PARENT_ELEMENT(this.props.componentId).getBoundingClientRect().left;
     const cursorTime = getTimeAtPixel(
-      x - leftOffset,
+      x - this.getGanttLeftOffset(),
       this.getStartDate(),
       this.getEndDate(),
       this.getTimelineWidth(),