+++ CHANGELOG.md
@@ -1,8 +1,24 @@
 ## Unreleased
 
+## v2.7.1
+
+* Updated `@famiprog-foundation/scriptable-ui` @ v1.0.0.
+
+## v2.7.0
+
+* [Zoom](https://github.com/flower-platform/react-timeline-10000/pull/58) and [small fix for Firefox](https://github.com/flower-platform/react-timeline-10000/pull/80)
+
+## v2.6.1
+
+* [Error when moving the mouse and the grid inside the gantt in not yet rendered](https://github.com/flower-platform/react-timeline-10000/pull/77)
+
+## v2.6.0
+
 * Internal: for `demo-app`, upgraded storybook version to v7.5. It also uses vite now.
 * Small issue: on props changed, the width of the table was reset to the original value.
 * [Highlighted interval positioned at the beginning of the gantt was not displayed](https://github.com/flower-platform/react-timeline-10000/pull/70)
+* [Controlled/uncontrolled split + rename `onTableResize` into `onSplitChange`](https://github.com/flower-platform/react-timeline-10000/pull/75)
+* [Add `displayItemOnSeparateRowIfOverlap` property](https://github.com/flower-platform/react-timeline-10000/pull/72/files)
 
 ## v2.5.0